requirements.txt
streamlit
mysql-connector-python
pandas
bcrypt
matplotlib 
